analysis/requirements.txt
affine==2.4.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
numpy==2.1.3
pocketbase==0.12.3
pyparsing==3.2.0
python-dotenv==1.0.1
rasterio==1.4.2
sniffio==1.3.1
